pyproject.toml
[tool.poetry]
name = "mcp-python-skeleton"
version = "0.1.0"
description = "Example MCP skeleton"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "mcpskeleton.__main__:main"

[tool.black]
extend_exclude = '.*_pb2.py|.*_pb2_grpc.py'

[tool.isort]
profile = 'black'
force_sort_within_sections = true
skip_gitignore = true
skip_glob = ['*_pb2.py', '*_pb2_grpc.py']

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uvicorn = {extras = ["standard"], version = "*"}
fastapi = ">=0.115.13"
pydantic-settings = ">=2.9.1"
mcp = {extras = ["mcp"], version = "*"}
fastmcp = "^2.10.6"
click = "*"


[tool.poetry.group.test.dependencies]
tox = "^4.14.2"
pre-commit = ">=4.2.0"
pytest = ">=8.4.1"
pytest-cov = ">=6.2.1"
pytest-asyncio = ">=1.0.0"
